+++ alchemy/src/docker/index.ts
@@ -1,4 +1,4 @@
 export * from "./api.js";
-export * from "./image.js";
+export * from "./remote-image.js";
 export * from "./container.js";
 export * from "./network.js";
