+++ alchemy/src/docker/remote-image.ts
