+++ examples/docker/.gitignore
