+++ examples/docker/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "docker-example",
+  "name": "docker",
   "version": "0.0.0",
   "private": true,
   "type": "module",
