+++ examples/docker/tsconfig.json
